% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{estimate}
\alias{estimate}
\title{Calculate silhouette coefficient and external metrics (ARI、NMI、Jaccard & Purity) for clustering based on true labels.}
\usage{
estimate(sc, true_labels)
}
\arguments{
\item{sc}{A Seurat object.}

\item{true_labels}{Cell type or morphology labels, etc., must be equal in number to the cells or spots in /{sc}.}
}
\value{
A list.
}
\description{
Calculate silhouette coefficient and external metrics (ARI、NMI、Jaccard & Purity) for clustering based on true labels.
}
