% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{featureSelect}
\alias{featureSelect}
\title{Screening for stable characteristic genes.}
\usage{
featureSelect(perf, stable = 0.9, org = NULL)
}
\arguments{
\item{perf}{Model cross-validation results output by sPLSDA.}

\item{stable}{Frequency of feature genes being selected in cross-validation.}

\item{org}{Corresponding species whole genome annotation.}
}
\value{
A matrix containing characteristic genes and corresponding stabilities.
}
\description{
Screening for stable characteristic genes.
}
