% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{makeCluster}
\alias{makeCluster}
\title{Standard workflow for dimensionality reduction and clustering of the RUD matrix/gene express matrix using the Seurat package.}
\usage{
makeCluster(
  seurat,
  norm = T,
  res = 0.1,
  dims = 1:50,
  nfeatures = 2000,
  k.param = 30,
  algorithm = 1,
  normalization.method = "LogNormalize"
)
}
\arguments{
\item{seurat}{A Seurat object.}

\item{norm}{Whether to normalize the matrix in the assay.}

\item{res}{Value of the resolution parameter in the \code{FindClusters()} function of the \code{Seurat} package.}

\item{dims}{Dimensions of reduction to use as input in the \code{FindNeighbors()} function of the \code{Seurat} package.}

\item{nfeatures}{Number of features to select as top variable features in the \code{FindVariableFeatures()} function of the \code{Seurat} package.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm in the \code{FindNeighbors()} function of the \code{Seurat} package.}

\item{algorithm}{Algorithm for modularity optimization in the \code{FindClusters()} function of the \code{Seurat} package.}

\item{normalization.method}{Method for normalization in the \code{NormalizeData()} function of the \code{Seurat} package.}
}
\value{
A Seurat object that has been reduced dimensionality using PCA and UMAP.
}
\description{
Standard workflow for dimensionality reduction and clustering of the RUD matrix/gene express matrix using the Seurat package.
}
