% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{PlotComppos}
\alias{PlotComppos}
\title{Plot the spatial distribution of component scores.}
\usage{
PlotComppos(sPLS, comps = "ALL", label, size = 0.45)
}
\arguments{
\item{sPLS}{A sPLS-DA model}

\item{comps}{A character vector containing the components to be involved in the plot, which must match the column names format of comp_score.}

\item{label}{A data frame, containing spatial coordinates (x, y), and labels (label).}

\item{size}{The size of the points.}
}
\value{
A list of ggplot objects, and prints a combined plot.
}
\description{
Plot the spatial distribution of component scores.
}
