% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{PlotCompumap}
\alias{PlotCompumap}
\title{Label positioning after dimensionality reduction}
\usage{
PlotCompumap(spls, size = 0.45)
}
\arguments{
\item{spls}{The final model output by sPLS-DA}

\item{size}{Size of the points in the plot.}
}
\value{
A ggplot object
}
\description{
Label positioning after dimensionality reduction
}
