% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{PlotCsub}
\alias{PlotCsub}
\title{Visualization of clusters.}
\usage{
PlotCsub(
  Seurat_Obj,
  type = c("celltype", "layer", "seurat_clusters"),
  subtype = NULL,
  size = 1
)
}
\arguments{
\item{Seurat_Obj}{Seurat object, the result of Seurat clustering, and the colored label is the corresponding annotation information.}

\item{type}{The label information used for coloring needs to be stored in the colData of the `seuratObject`.}

\item{subtype}{The annotation category to focus on, if not specified, select all.}

\item{size}{Scatter plot point size.}
}
\value{
A ggplot object
}
\description{
Visualization of clusters.
}
