% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{PlotLoad}
\alias{PlotLoad}
\title{Contribution of principal components}
\usage{
PlotLoad(
  GO_res,
  sPLS,
  gene_list = NULL,
  nterm = NULL,
  comp,
  org = "org.Hs.eg.db"
)
}
\arguments{
\item{GO_res}{Results of GO or other enrichment analysis.}

\item{sPLS}{The final discriminant model of `sPLSDA`.}

\item{gene_list}{Gene set of interest, if not specified, selects all genes.}

\item{nterm}{The number of terms that need to be displayed for enrichment analysis.}

\item{comp}{Principal components to be displayed}

\item{OrgDb}{Corresponding species whole genome annotation}
}
\value{
A bar chart
}
\description{
Contribution of principal components
}
