% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{sPLSDA}
\alias{sPLSDA}
\title{Run sPLS (sparse Partial Least Squares) supervised analysis.}
\usage{
sPLSDA(
  X,
  Y,
  ncomp,
  KeepX,
  folds = 10,
  nrepeat = 10,
  dist = "max.dist",
  measure = "BER",
  validation = "Mfold",
  invisible = T,
  nCore = 1
)
}
\arguments{
\item{X}{Matrix, rows are cells or spots, columns are genes.}

\item{Y}{Annotations of cells or spots corresponding to barcodes, a vector.}

\item{ncomp}{The expected number of components to be found.}

\item{KeepX}{The gene composition of each component.Grid of possible keepX values that will be tested for each comp}

\item{folds}{The number of folds for cross-validation in classification.}

\item{nrepeat}{The number of repetitions for cross-validation.For a thorough tuning step, the following code should be repeated 10 - 50 times and the error rate is averaged across the runs.}

\item{dist}{only applies to an object inheriting from "plsda" or "splsda" to evaluate the classification performance of the model. Should be a subset of "max.dist", "centroids.dist", "mahalanobis.dist".}

\item{measure}{Metrics for evaluating classification performance.}

\item{validation}{character. What kind of (internal) validation to use, matching one of "Mfold" or "loo". Default is "Mfold".}

\item{invisible}{If true, only the final model is output, otherwise the final model and classification verification results are output.}

\item{nCore}{multi-core computing.}
}
\value{
splsda classification model or a list containing the classification model and 10X cross-classification validation results
}
\description{
Run sPLS (sparse Partial Least Squares) supervised analysis.
}
