% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{ST11RUD}
\alias{ST11RUD}
\title{Gene Expression Matrix for ST11RUD}
\format{
A numeric matrix:
\describe{
  \item{Rows}{Genes, where each row corresponds to a specific gene.}
  \item{Columns}{Cells, where each column corresponds to a specific spot.}
}
}
\source{
Provided by the user for the ST11RUD experiment.
}
\usage{
ST11RUD
}
\description{
This dataset contains a RUD matrix where rows represent genes
and columns represent cells for the ST11RUDa experiment.
}
\examples{
data(ST11RUD)
dim(ST11RUD)  # View dimensions of the matrix
head(ST11RUD[, 1:5])  # View first 5 cells for the first few genes
}
\keyword{datasets}
