% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{WNNImpute}
\alias{WNNImpute}
\title{Imputation APA index using gene expression matrix and APA matrix.}
\usage{
WNNImpute(gene, APA, k = 20, init = TRUE)
}
\arguments{
\item{gene}{Reference gene expression matrix, with genes as rows and cells or spots as columns.}

\item{APA}{APA index to be recovered, with genes as rows and cells or spots as columns.}

\item{k}{Defines k for the k-nearest neighbor algorithm.}

\item{init}{Whether to initialize the APA index.}
}
\value{
Imputed APA index
}
\description{
A KNN-based and multimodal imputation model for recovering APA signals in the APA index matrix.
}
