% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spvAPA_funclib.R
\name{PlotCompumap}
\alias{PlotCompumap}
\title{Label positioning after dimensionality reduction}
\usage{
PlotCompumap(umap_df, Seurat_Obj, size, type = "label", subtype = NULL)
}
\arguments{
\item{umap_df}{A dataframe containing UMAP coordinates.}

\item{Seurat_Obj}{A Seurat object containing clustering information or other metadata.}

\item{size}{Size of the points in the plot.}

\item{type}{The type of label to use for coloring points. Default is "label".}

\item{subtype}{Optional argument specifying which subtypes (clusters) to highlight.}
}
\value{
A ggplot object
}
\description{
Label positioning after dimensionality reduction
}
